requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-environ==0.11.2
psycopg-binary==3.1.12
psycopg==3.1.12
flake8==6.1.0
Pillow==10.1.0